pyproject.toml
[tool.poetry]
name = "telegram-phone-number-checker"
version = "1.1.0"
description = "Check if phone numbers are connected to Telegram accounts."
authors = ["Bellingcat"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/bellingcat/telegram-phone-number-checker"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/bellingcat/telegram-phone-number-checker/issues"

[tool.poetry.scripts]
telegram-phone-number-checker = "telegram_phone_number_checker.main:main_entrypoint"

[tool.poetry.dependencies]
python = "^3.9"
anyio = "4.2.0"
click = "8.1.7"
exceptiongroup = "1.2.0"
idna = "3.6"
pyaes = "1.6.1"
pyasn1 = "0.5.1"
python-dotenv = "1.0.1"
rsa = "4.9"
sniffio = "1.3.0"
telethon = "1.34.0"
typing-extensions = "4.9.0"


[tool.poetry.group.dev.dependencies]
black = "24.3.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
